requirements.txt
openai
requests
supabase
python-dotenv